Add files via upload
Obtain the model the UR5 robot (Universal Robots): Draw the link frames and give the DH table of parameters, including the offsets so as to match the configuration shown in the figure. Create the corresponding SerialLink and plot it at configuration [0,0,0,0,0,0].

    ../_images/ur5_measures.png 

Recommended: You can do the deliverable with a Live Script using the MATLAB Live Editor, that can combine text and Matlab code (also, drawings of the frames can be done by hand and inserted in the Live Script as images).